Update variables.less and fix window resize bug
Change the breakpoint at which the website is resized for smaller screens. The previous size (1200px) was smaller than the actual width of the website, so there was a point at which the window size would leave elements of the website out, but the website would not resize since it was wider than 1200px. The new breakpoint makes sure that the website is resized before elements start to fall out.